Room Generation Updated
Added several essentials for map generation
- Rooms
- Room Generator
- Extra Info about TileSets

At this point you can use the Room generator to create a set of Rooms, but they won't be related to each other at all.

Added the ".*" (isDoor) specification to the map file format.

Added a toggle-able dev setting which shows which tiles are doors