Added Interlocking Rooms & Tile Orientation
The RoomManager will now, for each of the origin's doors, generate a room and line up one of that new room's doors to the origin's.
There is a small bug where rooms overlap or are a tile a way from the origin. However, the doors still line up.

In order to avoid having to have presets for each of a tiletype's rotations (e.g. Grass_TopLeft, Grass_Top, Grass_TopRight), tile rotation can now be specified in a file with "." + rotation:

.UL, .U, .UR, .R, .BR, .B, .BL, .L
U = upper, B = bottom, R - right, L = left

This will work as long as tilesets are passed in in the shape of the wall tileset.