Fixed Export Issue
Tileset was using the Content folder in the "Final Game" directory instead of the content folder in "net 6.0".

Whenever we tried to make a map using only what was in "net 6.0" Tileset couldn't find the necessary text files.

TL/DR: Tileset uses the txt files in the local net6.0/Content directory now :)